fix(types): resolve mypy errors and run mypy in CI
- Typed MODEL_REGISTRY with TypedDict for correct list[int] shapes
- Annotate capture/output streams and COM clients as Any for sounddevice/pycaw
- Fix AudioFormat.dtype return type for numpy dtype
- NPUEngine: guard active provider before accessing .value in load_model
- Spatial height filter: early-return when SOS/zi missing (runtime + typing)
- Separator: type NPU engine as Any
- Enhancer/noise: keep float32 arrays after scalar blend ops
- CI: run mypy src; pyproject mypy defaults

+++ src/audio/effects/enhancer.py
@@ -276,7 +276,7 @@ class AudioEnhancer:
         x = audio.astype(np.float32, copy=False)
         if use_npu:
             shaped = self._spectral_npu_ola(x)
-            x = (1.0 - blend) * x + blend * shaped
+            x = ((1.0 - blend) * x + blend * shaped).astype(np.float32, copy=False)
 
         # 1. Psychoacoustic bass
         if self.bass_boost > 0:
